Extract main menu button into separate fragment
