+++ app/src/main/res/layout/fragment_main_menu_button.xml
@@ -1,0 +1,16 @@
+<?xml version="1.0" encoding="utf-8"?>
+<androidx.constraintlayout.widget.ConstraintLayout
+    xmlns:android="http://schemas.android.com/apk/res/android"
+    android:layout_width="@dimen/map_button_size"
+    android:layout_height="@dimen/map_button_size">
+
+    <ImageButton
+        android:id="@+id/mainMenuButton"
+        style="@style/RoundWhiteButton"
+        android:layout_width="match_parent"
+        android:layout_height="match_parent"
+        android:contentDescription="@string/map_btn_menu"
+        android:scaleType="center"
+        android:src="@drawable/ic_menu_black_24dp" />
+
+</androidx.constraintlayout.widget.ConstraintLayout>
